fixed tests for signature list
